Again I can get both sim and pred. Need testing on database values and with code too. Now I need to include user preferences and also diff ranking for diff majors
